+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/topdata/node-agent/cmd"
+
+func main() {
+	cmd.Execute()
+}